refactor: UserInfoMapper가 UserService에 의존하지 않도록 변경
