+++ src/main/java/sixgaezzang/sidepeek/auth/oauth/service/strategy/UserInfoMapperFactory.java
@@ -4,16 +4,16 @@ import java.util.HashMap;
 import java.util.Map;
 import org.springframework.stereotype.Component;
 import sixgaezzang.sidepeek.auth.domain.ProviderType;
-import sixgaezzang.sidepeek.users.service.UserService;
+import sixgaezzang.sidepeek.auth.oauth.service.strategy.github.GithubUserInfoMapper;
 
 @Component
 public class UserInfoMapperFactory {
 
-    private Map<ProviderType, UserInfoMapper> mappers;
+    private final Map<ProviderType, UserInfoMapper> mappers;
 
-    public UserInfoMapperFactory(UserService userService) {
+    public UserInfoMapperFactory() {
         mappers = new HashMap<>();
-        mappers.put(ProviderType.GITHUB, new GithubUserInfoMapper(userService));
+        mappers.put(ProviderType.GITHUB, new GithubUserInfoMapper());
     }
 
     public UserInfoMapper getMapper(ProviderType providerType) {
